Safari-611.1.10/LayoutTests/http/tests/contentextensions/script-onerror-expected.txt

CONSOLE MESSAGE: Content blocker prevented frame displaying http://127.0.0.1:8000/contentextensions/script-onerror.html from loading a resource from http://127.0.0.1:8000/contentextensions/should-be-blocked.js
CONSOLE MESSAGE: Done executing script.  Waiting for onerror callback.
CONSOLE MESSAGE: onerror was correctly called on the next runloop cycle.
